requirements.txt
djangorestframework
Django==2.2.2
django-app-settings==0.5.1
django-braces==1.13.0
-e git://github.com/darklow/django-suit.git@v2#egg=django_suit
django-jet
six==1.12.0
sqlparse==0.3.0
mysqlclient
django-rest-auth
coreapi
django-coreapi
django-smtp-ssl
django-filter
django-url-filter